add tmux configuration
tmux is a software application that can be used to multiplex several virtual
consoles, allowing a user to access multiple separate terminal sessions inside
a single terminal window or remote terminal session. It is useful for dealing
with multiple programs from a command-line interface, and for separating
programs from the Unix shell that started the program.

More information can be found here: https://tmux.github.io/