add neovim configuration
Neovim is a refactor of Vim, that strives to be a superset of Vim. Neovim
shares the same configuration syntax with Vim; as a result, the same config
file can be used with both editors. Neovim is compatible with almost all of
Vim's features.

More information can be found here: https://neovim.io/