add lftp configuration
LFTP is a sophisticated ftp/http client, and a file transfer program supporting
a number of network protocols. Like BASH, it has job control and uses the
readline library for input. It has bookmarks, a built-in mirror command, and
can transfer several files in parallel. It was designed with reliability in
mind.

More information can be found here: https://lftp.yar.ru/