add zsh configuration
The Z shell (zsh) is a Unix shell that can be used as an interactive login
shell and as a powerful command interpreter for shell scripting. Zsh can be
thought of as an extended Bourne shell with a large number of improvements,
including some features of bash, ksh, and tcsh.

More information can be found here: http://www.zsh.org/