add GNU stow ignore file
GNU Stow is a symlink farm manager which takes distinct packages of software
and/or data located in separate directories on the filesystem, and makes
them appear to be installed in the same place. For example, /usr/local/bin
could contain symlinks to files within /usr/local/stow/emacs/bin,
/usr/local/stow/perl/bin etc., and likewise recursively for any other
subdirectories such as .../share, .../man, and so on.

In many situations, there will exist files under the package directories
which it would be undesirable to stow into the target directory. For
example, files related version control such as .gitignore, CVS, *,
v (RCS files) should typically not have symlinks from the target tree
pointing to them. Also there may be files or directories relating to the
build of the package which are not needed at run-time.

More information can be found here: https://www.gnu.org/software/stow/